1. Grid编辑器的列，应按schema中定义attribute的顺序排列 2. Grid编辑器增加新增、删除按钮 3. Grid编辑器与Tree同步显示 4. array显示为 [] 5. 增加xml文本编辑器 6. 属性编辑器的列宽，自动适应编辑器宽度 7.添加鼠标删除相应时间 8.去掉aurora命名

+++ uncertain_ide/uncertain/ide/eclipse/editor/PropertySorter.java
@@ -36,8 +36,7 @@ public class PropertySorter extends ViewerSorter {
 		int side = (order == ASCENDING) ? 1 : -1;
 		AttributeValue av = (AttributeValue) element;
 		if (element instanceof CategoryLabel) {
-			result = (Integer) PropertySheetContentProvider.Categorys.get(av
-					.getValueString());
+			result = ((Integer)PropertySheetContentProvider.Categorys.get(av.getValueString())).intValue();
 			result = result - side;
 		} else {
 			Category category = av.getAttribute().getCategoryInstance();
@@ -47,7 +46,7 @@ public class PropertySorter extends ViewerSorter {
 			} else {
 				cln = "Î´·Ö×é";
 			}
-			result = (Integer) PropertySheetContentProvider.Categorys.get(cln);
+			result = ((Integer) PropertySheetContentProvider.Categorys.get(cln)).intValue();
 			// int result=((File) element).isDirectory() ? 0 : 1;
 		}
 		if (order == DESCENDING)
